code is now better and bad idea is gone

+++ src/com/kport/langueg/lex/Token.java
@@ -4,7 +4,7 @@ public class Token {
     public TokenType tok;
     public String val;
 
-    public int lineNum;
+    public long lineNum;
 
     public Token(TokenType tok_){
         tok = tok_;
